pyproject.toml
[project]
name = "paper-digest"
version = "0.1.0"
description = "Automated paper digest using arXiv and Claude"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "feedparser>=6.0.0",
    "pyyaml>=6.0",
    "markdown>=3.5.0",
]

[project.scripts]
paper-digest = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
